refactor godoc out of github indexer
