+++ scheduler/indexer/awesome/fetch_awesome_go_list.go
@@ -1,4 +1,4 @@
-package github
+package awesome
 
 import (
 	"bytes"
