+++ scheduler/indexer/awesome/persist_awesome_packages.go
